(hel-537) settings routes
