+++ src/frontend/ui/parametrage-utilisateurs/usersListPage.tsx
@@ -1,0 +1,10 @@
+import "@gouvfr/dsfr/dist/component/table/table.min.css";
+
+export const UsersListPage = () => {
+
+    return (
+        <main className="fr-container">
+            page users
+        </main >
+    );
+};
