+++ src/pages/settings/profiles/[profileCode].tsx
@@ -1,9 +1,9 @@
 import { GetStaticPathsResult, GetStaticPropsResult } from "next";
 
-import { ProfileValue } from "../../../database/models/ProfilModel";
-import { getProfileByCodeEndpoint } from "../../backend/infrastructure/controllers/getProfileByCodeEndpoint";
-import { dependencies } from "../../backend/infrastructure/dependencies";
-import { ParametrageProfilPage } from "../../frontend/ui/parametrage-profil/ParametrageProfilPage";
+import { ProfileValue } from "../../../../database/models/ProfilModel";
+import { getProfileByCodeEndpoint } from "../../../backend/infrastructure/controllers/getProfileByCodeEndpoint";
+import { dependencies } from "../../../backend/infrastructure/dependencies";
+import { ParametrageProfilPage } from "../../../frontend/ui/parametrage-profil/ParametrageProfilPage";
 
 type RouterProps = Readonly<{
     profileValue: ProfileValue;
